requirements.txt
# Zerberus Pro 4.0 - Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.1
pyyaml==6.0
pytz==2023.3
aiofiles==23.2.1

# Datenbank
sqlalchemy==2.0.23
aiosqlite==0.19.0
# Sentiment (Patch 57: German BERT – VADER entfernt)
torch>=2.0.0
transformers>=4.35.0
sentencepiece>=0.1.99

# Overnight-Scheduler (Patch 57)
apscheduler>=3.10.0

# RAG
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# MQTT (optional)
paho-mqtt==1.6.1

# Telegram
python-telegram-bot==20.6

# WhatsApp (über Twilio)
twilio==8.10.0

# Profil-Authentifizierung
bcrypt==4.1.2
PyJWT==2.8.0

# RAG-Dokument-Upload (Patch 56)
python-docx==1.1.2
